Melhorias na componentização
Melhorias na componentização dos elementos do React. ListScreen foi subdividido em outros 4 elementos distintos. Pequenas mudanças na parte de estilização.
