+++ client/src/components/FilterInput.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function FilterInput( { searchText, onChangeSearchText } ) {
+    return (
+        <div>
+            <input type='text' placeholder='Filtro' value={searchText} onChange={onChangeSearchText}/>
+        </div>
+    );
+}
